Refactor: state machine without xstate (#33)
