+++ package.json
@@ -8,13 +8,11 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.3.0",
     "@types/jest": "^27.0.1",
-    "@types/lodash": "^4.14.182",
     "@types/node": "^18.0.6",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "cypress": "^10.3.1",
     "cypress-visual-regression": "^1.7.0",
-    "lodash": "^4.17.21",
     "perfect-freehand": "^1.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
