+++ src/CanvasForRect.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react'
 import { useState } from 'react'
+import { flushSync } from 'react-dom'
 import rough from 'roughjs/bundled/rough.esm'
 import { TCommitNewSnapshotParam, TElementData, TReplaceCurrentSnapshotParam } from './App'
 import { CONFIG } from './config'
-import { singletonThrottle } from './helpers/throttle'
 
 const generator = rough.generator({ options: { seed: CONFIG.SEED } })
 
@@ -92,9 +92,9 @@ export function CanvasForRect({
 
     // should come from onPointerDown()
     if (uiState.state === 'initDraw') {
-      // wrap in throttle because the following code need to be called at most once
+      // wrap in flushSync because the following code need to be called at most once
       // https://github.com/pobch/react-diagram/issues/27
-      singletonThrottle(() => {
+      flushSync(() => {
         const { sceneX, sceneY } = viewportCoordsToSceneCoords({
           viewportX: e.clientX,
           viewportY: e.clientY,
@@ -152,9 +152,6 @@ export function CanvasForRect({
     }
     // should come from onPointerMove()
     if (uiState.state === 'drawing') {
-      // reset the throttle timer that comes from the previous state's onPointerMove()
-      singletonThrottle.cancel()
-
       // adjust coord when finish drawing
       const drawnElement = getElementInCurrentSnapshot(uiState.data.elementId)
       if (!drawnElement || drawnElement.type !== 'rectangle') {
